@@ -1,36 +1,25 @@
 {
-  "name": "test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
+  "name": "about-me",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
-    "dev": "webpack-dev-server --hot --config ./build/webpack.dev.config.js",
-    "build": "webpack --config ./build/webpack.build.config.js"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build"
   },
-  "author": "xii",
-  "license": "ISC",
   "dependencies": {
-    "@babel/core": "^7.11.4",
-    "@mdi/font": "^5.5.55",
-    "babel-loader": "^8.1.0",
-    "css-loader": "^4.2.1",
-    "file-loader": "^6.0.0",
-    "mini-css-extract-plugin": "^0.10.0",
-    "moment": "^2.27.0",
-    "sass": "^1.26.10",
-    "style-loader": "^1.2.1",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.3",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.12",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "@mdi/font": "^5.8.55",
+    "axios": "^0.21.1",
+    "core-js": "^3.6.5",
+    "moment": "^2.29.1",
+    "vue": "^3.0.0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "copy-webpack-plugin": "^6.0.3",
-    "html-webpack-plugin": "^4.3.0",
-    "sass-loader": "^9.0.3",
-    "webpack-merge": "^5.1.2"
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-vuex": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^8.0.2"
   }
 }
